Remove test remove strategy to test_utils
