+++ tests/test_peaks.py
@@ -72,90 +72,3 @@ class TestComputePeakBasics(unittest.TestCase):
         test_data['dt'] = 1
         test_data['length'] = length
         return test_data
-
-
-def create_unique_intervals(size, time_range=(0, 40), allow_zero_length=True):
-    """
-    Hypothesis stragtegy which creates unqiue time intervals.
-
-    :param size: Number of intervals desired. Can be less if non-unique
-        intervals are found.
-    :param time_range: Time range in which intervals should be.
-    :param allow_zero_length: If true allow zero length intervals.
-    """
-    strat = strategies.lists(elements=strategies.integers(*time_range),
-                             min_size=size * 2,
-                             max_size=size * 2
-                             ).map(lambda x: _convert_to_interval(x, allow_zero_length))
-    return strat
-
-
-def _convert_to_interval(time_stamps, allow_zero_length):
-    time_stamps = np.sort(time_stamps)
-    intervals = np.zeros(len(time_stamps) // 2, strax.time_dt_fields)
-    intervals['dt'] = 1
-    intervals['time'] = time_stamps[::2]
-    intervals['length'] = time_stamps[1::2] - time_stamps[::2]
-
-    if not allow_zero_length:
-        intervals = intervals[intervals['length'] > 0]
-    return np.unique(intervals)
-
-
-@settings(deadline=None)
-@given(create_unique_intervals(10, time_range=(0, 30), allow_zero_length=False),
-       create_unique_intervals(5, time_range=(5, 25), allow_zero_length=False)
-       )
-def test_tag_peaks(peaks, veto_intervals):
-    peaks_in_vetos = strax.touching_windows(peaks, veto_intervals)
-
-    tags = np.zeros(len(peaks))
-    straxen.plugins.peak_processing.tag_peaks(tags, peaks_in_vetos, 1)
-
-    # Make an additional dummy array to test if function worked:
-    dtype = []
-    dtype += strax.time_dt_fields
-    dtype += [(('peak tag', 'tag'), np.int8)]
-    tagged_peaks = np.zeros(len(peaks), dtype)
-    tagged_peaks['time'] = peaks['time']
-    tagged_peaks['length'] = peaks['length']
-    tagged_peaks['dt'] = 1
-    tagged_peaks['tag'] = tags
-
-    split_tagged_peaks = strax.split_touching_windows(tagged_peaks, veto_intervals)
-
-    for split_peaks in split_tagged_peaks:
-        if not len(split_peaks):
-            continue
-        assert np.all(split_peaks['tag'] == 1), f'Not all peaks were tagged properly {split_peaks}'
-
-
-@settings(deadline=None)
-@given(create_unique_intervals(10, time_range=(50, 80), allow_zero_length=False),
-       create_unique_intervals(5, time_range=(55, 75), allow_zero_length=False)
-       )
-def test_get_time_to_clostest_veto(peaks, veto_intervals):
-    time_differences = straxen.plugins.peak_processing.get_time_to_closest_veto(peaks,
-                                                                                veto_intervals
-                                                                                )
-
-    if not len(peaks):
-        assert not len(time_differences), 'Input is empty but output is not?'
-
-    for ind, p in enumerate(peaks):
-        if len(veto_intervals):
-            dt = np.concatenate([veto_intervals['time'] - p['time'],
-                                 strax.endtime(veto_intervals) - p['time']])
-            # If the distance to the curren/next veto interval is identical
-            # the function favors positive values. Hence sort and reverse
-            # order for the test.
-            dt = np.sort(dt)[::-1]
-            ind_argmin = np.argmin(np.abs(dt))
-            dt = dt[ind_argmin]
-        else:
-            # If there are not any veto intervalls function will compute clostest
-            # difference to +/- 64 bit infinity.
-            dt = np.abs(straxen.INFINITY_64BIT_SIGNED - p['time'])
-
-        mes = f'Wrong time difference to closest event. expected: "{dt}" got: "{time_differences[ind]}"'
-        assert dt == time_differences[ind], mes
